+++ Relentless/src/Mesh/MeshFactory.cpp
@@ -11,8 +11,8 @@ namespace Relentless
 		meshes.clear();
 
 		Assimp::Importer importer;
-		constexpr const uint32_t flags = aiProcess_ConvertToLeftHanded | aiProcess_GenSmoothNormals | aiProcess_Triangulate
-			| aiProcess_ImproveCacheLocality | aiProcess_JoinIdenticalVertices | aiProcess_CalcTangentSpace | aiProcess_GenBoundingBoxes;
+		constexpr const uint32_t flags = (uint32_t)(aiProcess_ConvertToLeftHanded | aiProcess_GenSmoothNormals | aiProcess_Triangulate
+			| aiProcess_ImproveCacheLocality | aiProcess_JoinIdenticalVertices | aiProcess_CalcTangentSpace | aiProcess_GenBoundingBoxes);
 		const aiScene* pScene = importer.ReadFile(filePath.string(), flags);
 		RLS_ASSERT(pScene && !(pScene->mFlags & AI_SCENE_FLAGS_INCOMPLETE) && pScene->mRootNode, importer.GetErrorString());
